Lifted functions and fixed elements not using MUI
